refactor: rename port/integer.h -> util/integer.h
The functions in the integer module no longer match the standard C++
functions like from_chars. Move the integer module out of the port
module to avoid confusion.
