+++ src/quick-lint-js/cli/options.cpp
@@ -10,8 +10,8 @@
 #include <quick-lint-js/cli/options.h>
 #include <quick-lint-js/container/string-view.h>
 #include <quick-lint-js/io/output-stream.h>
-#include <quick-lint-js/port/integer.h>
 #include <quick-lint-js/port/warning.h>
+#include <quick-lint-js/util/integer.h>
 #include <quick-lint-js/util/narrow-cast.h>
 #include <string_view>
 #include <vector>
