+++ src/quick-lint-js/container/byte-buffer.h
@@ -9,7 +9,7 @@
 #include <quick-lint-js/assert.h>
 #include <quick-lint-js/port/char8.h>
 #include <quick-lint-js/port/have.h>
-#include <quick-lint-js/port/integer.h>
+#include <quick-lint-js/util/integer.h>
 #include <quick-lint-js/util/narrow-cast.h>
 #include <utility>
 #include <vector>
