+++ src/quick-lint-js/util/integer.cpp
@@ -10,8 +10,8 @@
 #include <cstdlib>
 #include <quick-lint-js/port/char8.h>
 #include <quick-lint-js/port/have.h>
-#include <quick-lint-js/port/integer.h>
 #include <quick-lint-js/port/warning.h>
+#include <quick-lint-js/util/integer.h>
 #include <quick-lint-js/util/math-overflow.h>
 #include <quick-lint-js/util/narrow-cast.h>
 #include <string>
