+++ src/quick-lint-js/util/integer.h
@@ -1,8 +1,8 @@
 // Copyright (C) 2020  Matthew "strager" Glazar
 // See end of file for extended copyright information.
 
-#ifndef QUICK_LINT_JS_PORT_INTEGER_H
-#define QUICK_LINT_JS_PORT_INTEGER_H
+#ifndef QUICK_LINT_JS_UTIL_INTEGER_H
+#define QUICK_LINT_JS_UTIL_INTEGER_H
 
 #include <cstddef>
 #include <limits>
