+++ test/test-integer-decimal.cpp
@@ -4,8 +4,8 @@
 #include <cstddef>
 #include <cstring>
 #include <gtest/gtest.h>
-#include <quick-lint-js/port/integer.h>
 #include <quick-lint-js/port/warning.h>
+#include <quick-lint-js/util/integer.h>
 #include <string_view>
 #include <system_error>
